Move `support logs` to `admin logs`
