+++ cmd/admin-console.go
@@ -47,12 +47,12 @@ var adminConsoleCmd = cli.Command{
 	Flags:              append(adminConsoleFlags, globalFlags...),
 	Hidden:             true,
 	HideHelpCommand:    true,
-	CustomHelpTemplate: "This command is not supported now and replaced by 'support logs show' command. Please use 'mc support logs show'.\n",
+	CustomHelpTemplate: "This command is not supported now and replaced by 'admin logs' command. Please use 'mc admin logs'.\n",
 }
 
 // mainAdminConsole - the entry function of console command
 func mainAdminConsole(ctx *cli.Context) error {
-	newCmd := []string{"mc support logs show"}
+	newCmd := []string{"mc admin logs"}
 
 	var flgStr string
